Extract other file for request in app review
