+++ lib/domain/record/components/button/taken_button.dart
@@ -5,7 +5,7 @@ import 'package:pilll/analytics.dart';
 import 'package:pilll/components/atoms/buttons.dart';
 import 'package:pilll/domain/modal/release_note.dart';
 import 'package:pilll/domain/record/record_page_state_notifier.dart';
-import 'package:pilll/domain/record/util/take.dart';
+import 'package:pilll/domain/record/util/request_in_app_review.dart';
 import 'package:pilll/entity/pill_sheet.codegen.dart';
 import 'package:pilll/entity/pill_sheet_group.codegen.dart';
 import 'package:pilll/error/error_alert.dart';
