+++ lib/domain/record/util/request_in_app_review.dart
@@ -1,0 +1,21 @@
+import 'package:in_app_review/in_app_review.dart';
+import 'package:pilll/util/shared_preference/keys.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+
+void requestInAppReview() {
+  SharedPreferences.getInstance().then((store) async {
+    final key = IntKey.totalCountOfActionForTakenPill;
+    int? value = store.getInt(key);
+    if (value == null) {
+      value = 0;
+    }
+    value += 1;
+    store.setInt(key, value);
+    if (value % 28 != 0) {
+      return;
+    }
+    if (await InAppReview.instance.isAvailable()) {
+      await InAppReview.instance.requestReview();
+    }
+  });
+}
