+++ lib/domain/record/util/take.dart
@@ -1,30 +1,9 @@
-import 'package:in_app_review/in_app_review.dart';
 import 'package:pilll/database/batch.dart';
 import 'package:pilll/entity/pill_sheet.codegen.dart';
 import 'package:pilll/entity/pill_sheet_group.codegen.dart';
 import 'package:pilll/database/pill_sheet.dart';
 import 'package:pilll/database/pill_sheet_group.dart';
 import 'package:pilll/database/pill_sheet_modified_history.dart';
-import 'package:pilll/util/shared_preference/keys.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-
-void requestInAppReview() {
-  SharedPreferences.getInstance().then((store) async {
-    final key = IntKey.totalCountOfActionForTakenPill;
-    int? value = store.getInt(key);
-    if (value == null) {
-      value = 0;
-    }
-    value += 1;
-    store.setInt(key, value);
-    if (value % 28 != 0) {
-      return;
-    }
-    if (await InAppReview.instance.isAvailable()) {
-      await InAppReview.instance.requestReview();
-    }
-  });
-}
 
 Future<PillSheetGroup?> take({
   required DateTime takenDate,
